packages/agent-tools/pyproject.toml
[project]
name = "agent-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "dongryeollee1", email = "cmeig21@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-experimental>=0.4.1",
    "langchain-tavily>=0.2.17",
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.22",
]

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"
